+++ src/data/use-cases/index.ts
@@ -83,7 +83,7 @@ export const pgNameMap: Record<string, string> = {
   'PG-SHP-APP-013': '项目会员页',
   'PG-SHP-APP-014': '直播详情页',
   'PG-OPS-PC-001': '搜索管理',
-  'PG-OPS-PC-002': '广告位管理',
+  'PG-OPS-PC-002': 'BANNER管理',
   'PG-OPS-PC-003': '金刚区管理',
   'PG-OPS-PC-004': '直播推荐管理',
   'PG-OPS-PC-005': '商品推荐管理',